Replace InterfaceSpy with adc.h (actual header) + adc.c (adc spy) + AdcSpy.h (adc spy header)
