+++ STM32/Firmware/RelayBypass/Source/Interface.c
@@ -1,6 +1,5 @@
 #include "../Include/Interface.h"
 //#include gpio.h
-//#include adc.h
 
 Status Interface_SwitchChannel(char channel) 
 {
@@ -147,4 +146,28 @@ Status Interface_ChangeRoute(char channel)
         break;
     }
     return OK;
+}
+
+LedColour Interface_GetChannel(char channel)
+{
+    uint32_t adcResult = 0;
+
+    switch (channel)
+    {
+    case CHANNEL_A:
+        // read ADC with HAL_ADC_GetValue(adc1);
+        adcResult = HAL_ADC_GetValue(&hadc1);
+        break;
+    case CHANNEL_B:
+        // read ADC with HAL_ADC_GetValue(adc2);
+        adcResult = HAL_ADC_GetValue(&hadc1);
+        break;
+    }
+
+    if (ADC_MIN_GREEN_BOUND <= adcResult)
+        return GREEN;
+    else if (ADC_MIN_RED_BOUND <= adcResult < ADC_MIN_GREEN_BOUND)
+        return RED;
+    else if (ADC_MIN_BLUE_BOUND <= adcResult < ADC_MIN_RED_BOUND)
+        return BLUE;
 }
