+++ STM32/Tests/TestsForStm32/ProductionCodeLib/ProductionCodeLib.vcxproj
@@ -155,6 +155,7 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClInclude Include="..\..\..\Firmware\RelayBypass\Core\Inc\adc.h" />
     <ClInclude Include="..\..\..\Firmware\RelayBypass\Core\Inc\gpio.h" />
     <ClInclude Include="..\..\..\Firmware\RelayBypass\Core\Inc\main.h" />
     <ClInclude Include="..\..\..\Firmware\RelayBypass\Include\Common.h" />
