+++ STM32/Tests/TestsForStm32/ProductionCodeLib/ProductionCodeLib.vcxproj.filters
@@ -39,6 +39,9 @@
     <ClInclude Include="..\..\..\Firmware\RelayBypass\Core\Inc\gpio.h">
       <Filter>HAL Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="..\..\..\Firmware\RelayBypass\Core\Inc\adc.h">
+      <Filter>HAL Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\..\..\Firmware\RelayBypass\Source\Serial.c">
