+++ STM32/Tests/TestsForStm32/TestsForStm32/SerialTests.cpp
@@ -4,7 +4,7 @@ extern "C"
 {
 #include "Serial.h"
 #include "SerialSpy.h"
-#include "InterfaceSpy.h"
+#include "adcSpy.h"
 }
 
 TEST_GROUP(Serial)
@@ -87,6 +87,7 @@ TEST(Serial, ShouldHandleGetCommand)
 {
     // Arrange – set command from PC tu UART Rx Buffer
     char getCmd[] = "get A\r\n";
+    hadc1.colour = RED;
     LONGS_EQUAL(OK, SerialSpy_SetRxBuffer(pSerial, getCmd, sizeof(getCmd)));
 
     // Act – Call to Handler for command processing
