+++ STM32/Tests/TestsForStm32/TestsForStm32/TestsForStm32.cpp
@@ -1,4 +1,5 @@
 #include "CppUTest/CommandLineTestRunner.h"
+
 int main(int argc, char** argv)
 {
     return RUN_ALL_TESTS(argc, argv);
