+++ STM32/Tests/TestsForStm32/TestsForStm32/gpio.c
@@ -31,3 +31,19 @@ Status USER_GPIO_HandOverLocalList(Node **pMasterList)
 
 	return OK;
 }
+
+void HAL_GPIO_WritePin(void* pGPIOx, uint16_t GPIO_Pin, GPIO_PinState PinState) 
+{
+	return;
+}
+
+void HAL_GPIO_TogglePin(void* pGPIOx, uint16_t GPIO_Pin)
+{
+	return;
+}
+
+//ToDo: Place in different location
+void HAL_Delay() 
+{
+	return;
+}
